LPS-82164 Move ServiceContextWrapper to providers package
