+++ modules/apps/headless-apio/forms/forms-apio-impl/src/main/java/com/liferay/forms/apio/internal/architect/provider/ServiceContextWrapper.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.forms.apio.internal;
+package com.liferay.forms.apio.internal.architect.provider;
 
 import com.liferay.portal.kernel.service.ServiceContext;
